refactor(main): diff viewer refresh helper を分離
refresh 系の Slint model 変換と history 更新を src/app/diff_viewer/refresh.rs に移し、main.rs の責務を callback 配線へ寄せる。

Refs: #224
