+++ src/app/diff_viewer/mod.rs
@@ -4,5 +4,6 @@
 //! 状態 / ヘルパは段階的にここへ移動する (#224 split)。
 
 pub(crate) mod snapshot;
+pub(crate) mod refresh;
 pub(crate) mod state;
 pub(crate) mod util;
